requirements.txt
fastapi
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
